+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * tsctl vs. timesketch-tools
 * fixed a small issue with beatifulsoup
 * update README.md
+* introduce sketch search
 
 # 2018-11-22